Display image and update Scaffold background color
The main function is updated to display an image from assets in the body of the Scaffold.
The Scaffold's background color is changed to deepOrangeAccent.
The pubspec.yaml file is updated to include the 'images/' directory in the assets.
An image file is renamed from 'Google-Flutter-Icon-PNG.jpg' to 'icon.jpg' within the 'images' directory.